src/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
flask-login
Werkzeug>=2.0.0
bcrypt
prometheus-client
python-json-logger
gunicorn